--- Code/Client/users/src/components/molecules/blog&newSection/blogsection2molecules/blogsection2molecule.module.css
@@ -1,93 +1,0 @@
-.blog_section{
-    display: flex;
-    flex-direction: column;
-}
-
-/* addOnsCard.module.css */
-/* 
-.heading {
-    display: flex;
-    flex-direction: column;
-    padding-left: 12px;
-    bottom: 1px solid ;
-}
-
-.h2{
-    font-size: 25px;
-}
-
-.p{
-    font-size: 18px;
-    font-weight: 500;
-}
-.Container {
-    display: flex;
-    flex-direction: column;
-}
-
-.addOnCard {
-    padding: 1rem 1.2rem;
-    border-radius: 1rem;
-    margin: 18px 12px;
-    height: 14.2rem;
-    background-color: white;
-}
-
-.cards{
-    display: flex;
-    flex-direction: row;
-}
-
-.cardTitle{
-    font-size: 19px;
-}
-
-.price{
-    display: flex;
-    justify-content: space-between;
-}
-
-.contanier{
-    width: 20rem;
-    padding-top: 6.3rem;
-}
-
-.card{
-    overflow: hidden;
-    padding: 1.5rem 1.8rem;
-    background-color: white;
-    box-shadow: 0 2px 8px 2px rgba(178, 178, 178, 0.189);
-  }
-
-.header{
-    border-bottom: 0.1rem solid rgb(182, 179, 179);
-    padding-bottom: 10px;
-}
-
-h2{
-    font-size: 21px;
-}
-
-.body{
-    padding-top: 10px;
-    border-bottom: 0.1rem solid rgb(182, 179, 179);
-    padding-bottom: 10px;
-}
-
-.icon{
-    font-size: 1.4rem;
-}
-
-.price{
-    padding-top: 10px;
-    display: flex;
-    justify-content: space-between;
-    font-size: 1.2rem;
-    font-weight: 700;
-}
-
-.btn{
-    display: flex;
-    justify-content: flex-end;
-    padding: 10px 0px;
-} */
