+++ Code/Client/users/src/components/sections/blog&new/blog&new1section/Blogs.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import style from './blogsection1molecule.module.css'
+import style from './blogs.module.css'
 
 export default function Blogsection1molecule() {
     return (
@@ -32,35 +32,7 @@ export default function Blogsection1molecule() {
                             <p>20 February / Global</p>
                         </div>
                     </div>
-                </div>
-
-                <div className={style.cards}>
-                <div className={style.blogCard}>
-                        <img src="" alt="blogImage" />
-                        <div>
-                            <p>Engineering, Backend, web</p>
-                            <h2>Building Scalable, Real-Time Chat to Improve Customer Experience</h2>
-                            <p>20 February / Global</p>
-                        </div>
-                    </div>
-
-                    <div className={style.blogCard}>
-                        <img src="" alt="blogImage" />
-                        <div>
-                            <p>Engineering, Backend, web</p>
-                            <h2>Building Scalable, Real-Time Chat to Improve Customer Experience</h2>
-                            <p>20 February / Global</p>
-                        </div>
-                    </div>
-
-                        <div className={style.blogCard}>
-                        <img src="" alt="blogImage" />
-                        <div>
-                            <p>Engineering, Backend, web</p>
-                            <h2>Building Scalable, Real-Time Chat to Improve Customer Experience</h2>
-                            <p>20 February / Global</p>
-                        </div>
-                    </div>
+                
                 </div>
             </div>
         </>
