+++ Code/Client/users/src/components/sections/blog&new/blog&new1section/blogs.module.css
@@ -20,8 +20,3 @@
     overflow: hidden;
     background-color: white;
 }
-
-.cards{
-    display: flex;
-    flex-direction: row;
-}
