+++ Code/Client/users/src/components/sections/blog&new/blog&new2section/blogfilter.module.css
@@ -1,0 +1,4 @@
+.blog_section{
+    display: flex;
+    flex-direction: column;
+}
